+++ CHANGELOG.md
@@ -20,6 +20,7 @@ master
 * Show an error message when git CLI is not available. #1765
 * Correctly show file names of GZIP'ed assets. #1364
 * Build file output is not parallel-ized! Use `middleman build --no-parallel` to disable.
+* Make template file extensions that get layouts by default configurable via `config[:extensions_with_layout]`
 
 # 4.0.0